Move Pages routing to App routing
